backend/requirements.txt
Django>=3.0,<4.0
djangorestframework==3.12.4
django-cors-headers==3.13.0
psycopg2==2.9.3
djoser==2.1.0
PyJWT==2.0.0
djangorestframework-simplejwt==4.6.0
